refactor(client): centralize map-like reply primitives in reply-utils
isPlainObject / mapLikeEntries / mapLikeValues / mapLikeToObject /
mapLikeToFlatArray / getMapValue were independently re-implemented in
@redis/client (HOTKEYS_GET), @redis/search (reply-transformers), and
@redis/time-series (INFO, INFO_DEBUG) with subtly different shapes —
some lacked null-safe key stringification, some over-accepted Buffer/
typed-array values as plain objects, and INFO_DEBUG carried a fourth
slightly divergent variant. Consolidate into a single module in
@redis/client and have all consumers import from it.

The unified versions adopt the strictest behavior already validated by
prior fixes: isPlainObject rejects Buffer/typed-arrays, mapLikeEntries
auto-unwraps single-element wrappers only when the inner value is
Array/Map/plain-object, and keyToString short-circuits null/undefined
keys. search/reply-transformers re-exports the shared symbols so
existing search command imports remain unchanged.
