+++ packages/search/lib/commands/reply-transformers.ts
@@ -1,40 +1,20 @@
-function isPlainObject(value: unknown): value is Record<string, unknown> {
-  return value !== null &&
-    typeof value === 'object' &&
-    !Array.isArray(value) &&
-    !(value instanceof Map);
-}
+import {
+  isPlainObject,
+  mapLikeEntries,
+  mapLikeValues,
+  mapLikeToObject,
+  mapLikeToFlatArray,
+  getMapValue
+} from '@redis/client/dist/lib/commands/reply-utils';
 
-export function mapLikeEntries(value: unknown): Array<[string, unknown]> {
-  if (value instanceof Map) {
-    return Array.from(value.entries(), ([key, entryValue]) => [key.toString(), entryValue]);
-  }
-
-  if (Array.isArray(value)) {
-    if (
-      value.length === 1 &&
-      (Array.isArray(value[0]) || value[0] instanceof Map || isPlainObject(value[0]))
-    ) {
-      return mapLikeEntries(value[0]);
-    }
-
-    if (value.every(item => Array.isArray(item) && item.length >= 2)) {
-      return value.map(item => [item[0].toString(), item[1]]);
-    }
-
-    const entries: Array<[string, unknown]> = [];
-    for (let i = 0; i < value.length - 1; i += 2) {
-      entries.push([value[i].toString(), value[i + 1]]);
-    }
-    return entries;
-  }
-
-  if (isPlainObject(value)) {
-    return Object.entries(value);
-  }
-
-  return [];
-}
+export {
+  isPlainObject,
+  mapLikeEntries,
+  mapLikeValues,
+  mapLikeToObject,
+  mapLikeToFlatArray,
+  getMapValue
+};
 
 export function toCompatObject(value: Record<string, unknown>): Record<string, unknown> {
   const descriptors: PropertyDescriptorMap = {};
@@ -49,56 +29,6 @@ export function toCompatObject(value: Record<string, unknown>): Record<string, u
   }
 
   return Object.defineProperties({}, descriptors);
-}
-
-export function mapLikeToObject(value: unknown): Record<string, unknown> {
-  const object: Record<string, unknown> = {};
-  for (const [key, entryValue] of mapLikeEntries(value)) {
-    object[key] = entryValue;
-  }
-  return object;
-}
-
-export function mapLikeToFlatArray(value: unknown): Array<unknown> {
-  const flat: Array<unknown> = [];
-  for (const [key, entryValue] of mapLikeEntries(value)) {
-    flat.push(key, entryValue);
-  }
-  return flat;
-}
-
-export function mapLikeValues(value: unknown): Array<unknown> {
-  if (Array.isArray(value)) return value;
-  if (value instanceof Map) return [...value.values()];
-  if (isPlainObject(value)) return Object.values(value);
-  return [];
-}
-
-export function getMapValue(value: unknown, keys: Array<string>): unknown {
-  const object = mapLikeToObject(value);
-
-  for (const key of keys) {
-    if (Object.hasOwn(object, key)) {
-      return object[key];
-    }
-  }
-
-  const lowerCaseKeyToOriginal = new Map<string, string>();
-  for (const key of Object.keys(object)) {
-    const lowerCaseKey = key.toLowerCase();
-    if (!lowerCaseKeyToOriginal.has(lowerCaseKey)) {
-      lowerCaseKeyToOriginal.set(lowerCaseKey, key);
-    }
-  }
-
-  for (const key of keys) {
-    const original = lowerCaseKeyToOriginal.get(key.toLowerCase());
-    if (original !== undefined) {
-      return object[original];
-    }
-  }
-
-  return undefined;
 }
 
 function assignDocumentField(target: Record<string, unknown>, key: string, value: unknown): void {
